+++ README.md
@@ -1,4 +1,4 @@
 # API Kit - Monorepo
-Jump to [package: api-kit](./packages/api-kit) 
+Jump to [package: api-kit](./packages/@api-kit) 
 
 Former `sveltekit-zero-api` — rewrite to include test suites, wrap into a monorepo, folloing template guidelines [refzlund/giraffe](https://github.com/refzlund/giraffe).
