+++ package.json
@@ -1,34 +1,38 @@
 {
-	"name": "api-kit-monorepo",
-	"version": "0.0.0",
-	"description": "A monorepo for api-kit.",
-	"private": true,
-	"type": "module",
-	"scripts": {
-		"test": "pnpm -r test",
-		"lint": "pnpm -r lint",
-		"typecheck": "pnpm -r typecheck",
-		"precommit": "pnpm -r precommit",
-		"build": "pnpm -r build",
-		"publish": "pnpm build && changeset publish"
-	},
-	"devDependencies": {
-		"@changesets/cli": "^2.26.2",
-		"@svitejs/changesets-changelog-github-compact": "^1.1.0",
-		"@types/node": "20.4.2",
-		"eslint": "^8.44.0",
-		"eslint-configs": "workspace:*",
-		"ts-configs": "workspace:*",
-		"vite-configs": "workspace:*",
-		"prettier": "^3.0.0",
-		"typescript": "^5.1.6",
-		"vite": "^4.3.9",
-		"vite-plugin-dts": "^3.1.0",
-		"vite-tsconfig-paths": "^4.2.0",
-		"vitest": "^0.33.0"
-	},
-	"packageManager": "pnpm@8.6.7",
-	"engines": {
-		"pnpm": "^8.0.0"
-	}
+  "name": "api-kit-monorepo",
+  "version": "0.0.0",
+  "description": "A monorepo for api-kit.",
+  "private": true,
+  "type": "module",
+  "workspaces": [
+    "packages/configs/*",
+    "packages/*"
+  ],
+  "scripts": {
+    "test": "pnpm -r test",
+    "lint": "pnpm -r lint",
+    "typecheck": "pnpm -r typecheck",
+    "precommit": "pnpm -r precommit",
+    "build": "pnpm -r build",
+    "publish": "pnpm build && changeset publish"
+  },
+  "devDependencies": {
+    "@changesets/cli": "^2.26.2",
+    "@svitejs/changesets-changelog-github-compact": "^1.1.0",
+    "@types/node": "20.4.2",
+    "eslint": "^8.44.0",
+    "eslint-configs": "workspace:*",
+    "prettier": "^3.0.0",
+    "ts-configs": "workspace:*",
+    "typescript": "^5.2.2",
+    "vite": "^4.3.9",
+    "vite-configs": "workspace:*",
+    "vite-plugin-dts": "^3.1.0",
+    "vite-tsconfig-paths": "^4.2.0",
+    "vitest": "^0.33.0"
+  },
+  "packageManager": "pnpm@8.6.7",
+  "engines": {
+    "pnpm": "^8.0.0"
+  }
 }
