+++ packages/@api-kit/.eslintrc.cjs
@@ -1,0 +1,6 @@
+// @ts-check
+/* eslint @typescript-eslint/no-var-requires: 0 */
+const { mergeConfigs, config } = require('eslint-configs')
+
+module.exports = mergeConfigs(config.ts, {
+})
