+++ packages/@api-kit/src/index.ts
@@ -1,0 +1,12 @@
+const str = `Giraffe's are quite intruiging creatures.`
+export default function () {
+	return str
+}
+
+function endpoint(){}
+
+
+export const GET = endpoint(
+
+)
+
