+++ packages/@api-kit/src/types/json.ts
@@ -1,0 +1,19 @@
+type JsonType = boolean | number | string | null | undefined
+
+type Jsonable = { toJSON(): unknown }
+
+type InferToJSON<T extends Jsonable> =
+	T extends { toJSON(): infer U } ? U extends object ? ToJSON<U> : U : never
+
+type Jsonify<T> =
+	T extends Function | symbol | undefined ? never :
+	T extends Jsonable ? InferToJSON<T> :
+	T extends JsonType ? T :
+	T extends object ? {
+		-readonly [K in keyof T as Jsonify<T[K]> extends never ? never : K]: Jsonify<T[K]>
+	} : never
+
+export type ToJSON<T extends object | undefined | unknown> =
+	T extends undefined | string | number | boolean | symbol | FormData | Blob
+	? T
+	: Jsonify<T>
