--- packages/api-kit/.drawio
@@ -1,27 +1,0 @@
-<mxfile host="65bd71144e">
-    <diagram id="fWfo3xHF7rSx2LhqYZcW" name="Page-1">
-        <mxGraphModel dx="1222" dy="924" grid="1" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="0" pageScale="1" pageWidth="850" pageHeight="1100" math="0" shadow="0">
-            <root>
-                <mxCell id="0"/>
-                <mxCell id="1" parent="0"/>
-                <mxCell id="10" style="edgeStyle=none;html=1;" edge="1" parent="1" source="3" target="9">
-                    <mxGeometry relative="1" as="geometry"/>
-                </mxCell>
-                <mxCell id="3" value="&lt;pre style=&quot;line-height: 20px;&quot;&gt;&lt;div style=&quot;&quot;&gt;&lt;font color=&quot;#e3e2e0&quot; face=&quot;FiraCode Nerd Font, Consolas, Courier New, monospace&quot;&gt;&lt;span style=&quot;font-size: 11px;&quot;&gt;&lt;br&gt;&lt;/span&gt;&lt;/font&gt;&lt;/div&gt;&lt;div style=&quot;color: rgb(227, 226, 224); font-family: &amp;quot;FiraCode Nerd Font&amp;quot;, Consolas, &amp;quot;Courier New&amp;quot;, monospace; font-size: 11px;&quot;&gt;&lt;span style=&quot;color: #edb92e;&quot;&gt;export&lt;/span&gt; &lt;span style=&quot;color: #ff4e50;&quot;&gt;const&lt;/span&gt; GET &lt;span style=&quot;color: #edb92e;&quot;&gt;=&lt;/span&gt; &lt;span style=&quot;color: #c3cb4c;&quot;&gt;endpoint&lt;/span&gt;(&lt;/div&gt;&lt;div style=&quot;&quot;&gt;&lt;span style=&quot;white-space: pre;&quot;&gt;&#9;&lt;/span&gt;&lt;br&gt;&lt;/div&gt;&lt;div style=&quot;color: rgb(227, 226, 224); font-family: &amp;quot;FiraCode Nerd Font&amp;quot;, Consolas, &amp;quot;Courier New&amp;quot;, monospace; font-size: 11px;&quot;&gt;)&lt;/div&gt;&lt;div style=&quot;&quot;&gt;&lt;br&gt;&lt;/div&gt;&lt;/pre&gt;" style="html=1;fillColor=#101010;strokeColor=#080808;align=left;verticalAlign=top;spacingTop=0;spacingLeft=0;spacing=20;whiteSpace=wrap;labelBorderColor=none;labelBackgroundColor=none;" vertex="1" parent="1">
-                    <mxGeometry x="90" y="70" width="330" height="180" as="geometry"/>
-                </mxCell>
-                <mxCell id="12" style="edgeStyle=none;html=1;" edge="1" parent="1" source="9">
-                    <mxGeometry relative="1" as="geometry">
-                        <mxPoint x="380" y="370" as="targetPoint"/>
-                    </mxGeometry>
-                </mxCell>
-                <mxCell id="9" value="TEndpoint" style="whiteSpace=wrap;html=1;" vertex="1" parent="1">
-                    <mxGeometry x="195" y="340" width="120" height="60" as="geometry"/>
-                </mxCell>
-                <mxCell id="13" value="TResponse" style="whiteSpace=wrap;html=1;" vertex="1" parent="1">
-                    <mxGeometry x="380" y="340" width="120" height="60" as="geometry"/>
-                </mxCell>
-            </root>
-        </mxGraphModel>
-    </diagram>
-</mxfile>
