--- packages/api-kit/CHANGELOG.md
@@ -1,7 +1,0 @@
-# api-kit
-
-## 0.1.1
-
-### Patch Changes
-
-- chore: initial configuration ([#3](https://github.com/Refzlund/api-kit/pull/3))
