--- packages/api-kit/package.json
@@ -1,43 +1,0 @@
-{
-	"name": "api-kit",
-	"version": "0.0.1",
-	"description": "Library template that uses Vite for building, Vitest for testing, Changesets for versioning, renovate for upgrading packages.",
-	"keywords": [],
-	"author": "Arthur Refslund @Refzlund",
-	"license": "MIT",
-	"repository": "github:refzlund/api-kit",
-	"homepage": "https://github.com/refzlund/api-kit",
-	"bugs": {
-		"url": "https://github.com/refzlund/api-kit/labels/bug"
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"scripts": {
-		"lint": "eslint **/*.{js,ts,cjs}",
-		"typecheck": "vitest typecheck -t \".*\\.ts\"",
-		"test": "vitest",
-		"precommit": "pnpm run lint && pnpm run typecheck && pnpm run test",
-		"build": "vite build"
-	},
-	"type": "module",
-	"main": "./.dist/index.js",
-	"module": "./.dist/index.js",
-	"types": "./.dist/index.d.ts",
-	"exports": {
-		".": {
-			"import": "./.dist/index.js",
-			"require": "./.dist/index.umd.cjs"
-		}
-	},
-	"files": [
-		".dist"
-	],
-	"devDependencies": {
-		"@sveltejs/kit": "^1.22.3"
-	},
-	"packageManager": "pnpm@8.6.7",
-	"engines": {
-		"pnpm": "^8.0.0"
-	}
-}
