--- packages/api-kit/src/index.ts
@@ -1,12 +1,0 @@
-const str = `Giraffe's are quite intruiging creatures.`
-export default function () {
-	return str
-}
-
-function endpoint(){}
-
-
-export const GET = endpoint(
-
-)
-
