--- packages/api-kit/tsconfig.json
@@ -1,19 +1,0 @@
-{
-	"extends": "ts-configs/tsconfig.def.json",
-	"compilerOptions": {
-		/* Paths */
-		"baseUrl": ".",
-		"paths": {
-			// Library
-			"api-kit": ["./.dist"],
-			"api-kit/*": ["./.dist/*"],
-			// ...
-			"$": ["./src"],
-			"$/*": ["./src/*"],
-
-			"$types": ["src/types"],
-			"$types/*": ["src/types/*"]
-		}
-	},
-	"include": ["src", "src/types", ".dist", "_examples"]
-}
